chore: move enum converters to separate namespace
