+++ FlameCsv.Benchmark/EnumFormatBench.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text.Unicode;
 using FlameCsv.Attributes;
 using FlameCsv.Converters;
+using FlameCsv.Converters.Enums;
 
 namespace FlameCsv.Benchmark;
