+++ FlameCsv.Core/Converters/Enums/CsvEnumFlagsTextFormatStrategy.cs
@@ -1,6 +1,6 @@
 ﻿using FlameCsv.Utilities;
 
-namespace FlameCsv.Converters;
+namespace FlameCsv.Converters.Enums;
 
 /// <summary>
 /// Internal implementation detail, used by the source generator.
