+++ FlameCsv.Core/Converters/Enums/EnumUtf8Converter.cs
@@ -1,12 +1,11 @@
 using System.Buffers;
 using System.Collections.Frozen;
-using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.Unicode;
 using FlameCsv.Extensions;
 using FlameCsv.Utilities;
 
-namespace FlameCsv.Converters;
+namespace FlameCsv.Converters.Enums;
 
 /// <summary>
 /// The default converter for non-flags enums.
