+++ FlameCsv.Tests/Converters/OldEnumTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Runtime.Serialization;
 using System.Text;
 using FlameCsv.Converters;
+using FlameCsv.Converters.Enums;
 using FlameCsv.Utilities;
 using FlameCsv.Utilities.Comparers;
